# docs/swagger.yaml
basePath: /api
definitions:
...
  asset.AssetType:
    enum:
    - CASH
    - TERM_DEPOSIT
    - STOCK
    - ETF
    - FUND
    - BOND
    - CRYPTOCURRENCY
    - NFT
    - DEFI_TOKEN
    - PRECIOUS_METAL
    - REAL_ESTATE
    - DEBT
    - RECEIVABLE
    - SALARY
    - OTHER
    type: string
    x-enum-varnames:
    - Cash
    - TermDeposit
    - Stock
    - ETF
    - Fund
    - Bond
    - Cryptocurrency
    - NFT
    - DeFiToken
    - PreciousMetal
    - RealEstate
    - Debt
    - Receivable
    - Salary
    - Other
...
  routes.AssetResponse:
    properties:
      accountId:
        type: string
      createdAt:
        type: string
      definitionId:
        type: string
      id:
        type: string
      notes:
        type: string
      purchaseDate:
        type: string
      quantity:
        type: number
      type:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  routes.CreateAssetRequest:
    properties:
      accountId:
        type: string
      definitionId:
        type: string
      notes:
        type: string
      purchaseDate:
        type: integer
      quantity:
        type: number
      type:
        $ref: '#/definitions/asset.AssetType'
    required:
    - accountId
    - definitionId
    - purchaseDate
    - quantity
    - type
    type: object
...
  routes.UpdateAssetRequest:
    properties:
      accountId:
        type: string
      definitionId:
        type: string
      notes:
        type: string
      purchaseDate:
        type: integer
      quantity:
        type: number
      type:
        $ref: '#/definitions/asset.AssetType'
    required:
    - accountId
    - definitionId
    - purchaseDate
    - quantity
    - type
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a wallet service API for managing user accounts, assets, and
    transactions
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Wallet Service API
  version: "1.0"
paths:
...
  /assets:
    post:
      consumes:
      - application/json
      description: Create a new asset for the authenticated user
      parameters:
      - description: Asset creation data
        in: body
        name: asset
        required: true
        schema:
          $ref: '#/definitions/routes.CreateAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AssetResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new asset
      tags:
      - assets
  /assets/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing asset for the authenticated user
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Asset update data
        in: body
        name: asset
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AssetResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an asset
      tags:
      - assets
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"